Move electron API out of renderer code
